# docs/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "FinSense AI API",
  "description": "Financial analytics, portfolio optimisation, ML forecasting and RAG research API. Decision support only - not investment advice.",
  "version": "1.0.0"
 },
 "paths": {
...
  "/api/v1/auth/register": {
   "post": {
    "tags": [
     "auth"
    ],
    "summary": "Register",
    "operationId": "register_api_v1_auth_register_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RegisterRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/UserOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/auth/login": {
   "post": {
    "tags": [
     "auth"
    ],
    "summary": "Login",
    "operationId": "login_api_v1_auth_login_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/LoginRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SessionOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/auth/me": {
   "get": {
    "tags": [
     "auth"
    ],
    "summary": "Me",
    "operationId": "me_api_v1_auth_me_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/UserOut"
        }
       }
      }
     }
    }
   },
   "patch": {
    "tags": [
     "auth"
    ],
    "summary": "Update Me",
    "operationId": "update_me_api_v1_auth_me_patch",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ProfileUpdate"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/UserOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/auth/change-password": {
   "post": {
    "tags": [
     "auth"
    ],
    "summary": "Change Password",
    "operationId": "change_password_api_v1_auth_change_password_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ChangePasswordRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/auth/password-reset/request": {
   "post": {
    "tags": [
     "auth"
    ],
    "summary": "Password Reset Request",
    "operationId": "password_reset_request_api_v1_auth_password_reset_request_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ResetRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "202": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response Password Reset Request Api V1 Auth Password Reset Request Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/auth/password-reset/confirm": {
   "post": {
    "tags": [
     "auth"
    ],
    "summary": "Password Reset Confirm",
    "operationId": "password_reset_confirm_api_v1_auth_password_reset_confirm_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ResetConfirm"
       }
      }
     },
     "required": true
    },
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/assistant/query": {
   "post": {
    "tags": [
     "assistant"
    ],
    "summary": "Query",
    "operationId": "query_api_v1_assistant_query_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/QueryRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response Query Api V1 Assistant Query Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/assistant/conversations/{conversation_id}": {
   "get": {
    "tags": [
     "assistant"
    ],
    "summary": "Get Conversation",
    "operationId": "get_conversation_api_v1_assistant_conversations__conversation_id__get",
    "parameters": [
     {
      "name": "conversation_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Conversation Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Get Conversation Api V1 Assistant Conversations  Conversation Id  Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "assistant"
    ],
    "summary": "Delete Conversation",
    "operationId": "delete_conversation_api_v1_assistant_conversations__conversation_id__delete",
    "parameters": [
     {
      "name": "conversation_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Conversation Id"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "ChangePasswordRequest": {
    "properties": {
     "current_password": {
      "type": "string",
      "maxLength": 128,
      "minLength": 1,
      "title": "Current Password"
     },
     "new_password": {
      "type": "string",
      "maxLength": 128,
      "minLength": 1,
      "title": "New Password"
     }
    },
    "type": "object",
    "required": [
     "current_password",
     "new_password"
    ],
    "title": "ChangePasswordRequest"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
   "LoginRequest": {
    "properties": {
     "email": {
      "type": "string",
      "format": "email",
      "title": "Email"
     },
     "password": {
      "type": "string",
      "maxLength": 128,
      "minLength": 1,
      "title": "Password"
     }
    },
    "type": "object",
    "required": [
     "email",
     "password"
    ],
    "title": "LoginRequest"
   },
...
   "Preferences": {
    "properties": {
     "risk_free_rate": {
      "anyOf": [
       {
        "type": "number",
        "maximum": 0.2,
        "minimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Risk Free Rate"
     },
     "default_period": {
      "anyOf": [
       {
        "type": "string",
        "enum": [
         "1y",
         "3y",
         "5y",
         "max"
        ]
       },
       {
        "type": "null"
       }
      ],
      "title": "Default Period"
     },
     "risk_profile": {
      "anyOf": [
       {
        "type": "string",
        "enum": [
         "conservative",
         "moderate",
         "aggressive"
        ]
       },
       {
        "type": "null"
       }
      ],
      "title": "Risk Profile"
     }
    },
    "type": "object",
    "title": "Preferences"
   },
   "ProfileUpdate": {
    "properties": {
     "display_name": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 80,
        "minLength": 1
       },
       {
        "type": "null"
       }
      ],
      "title": "Display Name"
     },
     "preferences": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/Preferences"
       },
       {
        "type": "null"
       }
      ]
     }
    },
    "type": "object",
    "title": "ProfileUpdate"
   },
   "QueryRequest": {
    "properties": {
     "question": {
      "type": "string",
      "maxLength": 1000,
      "minLength": 2,
      "title": "Question"
     },
     "conversation_id": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Conversation Id"
     },
     "document_ids": {
      "anyOf": [
       {
        "items": {
         "type": "string",
         "format": "uuid"
        },
        "type": "array",
        "maxItems": 50
       },
       {
        "type": "null"
       }
      ],
      "title": "Document Ids"
     }
    },
    "type": "object",
    "required": [
     "question"
    ],
    "title": "QueryRequest"
   },
   "RegisterRequest": {
    "properties": {
     "email": {
      "type": "string",
      "format": "email",
      "title": "Email"
     },
     "password": {
      "type": "string",
      "maxLength": 128,
      "minLength": 1,
      "title": "Password"
     },
     "display_name": {
      "type": "string",
      "maxLength": 80,
      "minLength": 1,
      "title": "Display Name"
     }
    },
    "type": "object",
    "required": [
     "email",
     "password",
     "display_name"
    ],
    "title": "RegisterRequest"
   },
   "ResetConfirm": {
    "properties": {
     "token": {
      "type": "string",
      "maxLength": 200,
      "minLength": 20,
      "title": "Token"
     },
     "new_password": {
      "type": "string",
      "maxLength": 128,
      "minLength": 1,
      "title": "New Password"
     }
    },
    "type": "object",
    "required": [
     "token",
     "new_password"
    ],
    "title": "ResetConfirm"
   },
   "ResetRequest": {
    "properties": {
     "email": {
      "type": "string",
      "format": "email",
      "title": "Email"
     }
    },
    "type": "object",
    "required": [
     "email"
    ],
    "title": "ResetRequest"
   },
...
   "SessionOut": {
    "properties": {
     "user": {
      "$ref": "#/components/schemas/UserOut"
     },
     "csrf_token": {
      "type": "string",
      "title": "Csrf Token"
     },
     "expires_at": {
      "type": "string",
      "format": "date-time",
      "title": "Expires At"
     }
    },
    "type": "object",
    "required": [
     "user",
     "csrf_token",
     "expires_at"
    ],
    "title": "SessionOut"
   },
...
   "UserOut": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "email": {
      "type": "string",
      "title": "Email"
     },
     "display_name": {
      "type": "string",
      "title": "Display Name"
     },
     "role": {
      "type": "string",
      "title": "Role"
     },
     "preferences": {
      "additionalProperties": true,
      "type": "object",
      "title": "Preferences"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     }
    },
    "type": "object",
    "required": [
     "id",
     "email",
     "display_name",
     "role",
     "preferences",
     "created_at"
    ],
    "title": "UserOut"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  }
 }
}